requirements.txt
numpy>=1.23.5
opencv-python>=4.10.0
PyYAML>=6.0.1
tqdm>=4.65.0

easyocr>=1.6.2

torch>=2.4.0
torchvision>=0.19.0
ultralytics>=8.0.43